mcp-server/requirements.txt
mcp>=0.9.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sse-starlette>=1.8.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
